Update useCases section
