+++ app/styles.css
@@ -165,206 +165,20 @@
   }
 }
 
-@utility card-stack {
-  --padding-offset: 60px;
-  --scale-pct: 0.1;
-}
+@utility border-linear-gradient {
+  --bg: #151513;
+  --radius: 16px;
+  --gradient:
+    118.71deg, #ffffff 11.96%, #212121 45.04%, #bdbdbd 66.69%, #232323 75.79%,
+    #646464 86.15%;
 
-@utility card-stack-item {
-  @apply absolute left-[calc(var(--index)*var(--padding-offset))] z-(--index) origin-left scale-100 scale-[calc(1-3*var(--scale-pct))] transition-all duration-200 ease-in-out;
+  @apply relative z-1 rounded-(--radius) border border-transparent bg-(--bg);
 
-  &:has(+ [class*="card-stack-item"]:hover) {
-    @apply !scale-[calc(1-var(--scale-pct))];
-  }
-  &:has(+ * + [class*="card-stack-item"]:hover) {
-    @apply !scale-[calc(1-2*var(--scale-pct))];
+  &::before {
+    @apply absolute inset-0 -z-1 rounded-(--radius) bg-linear-(--gradient) content-[''];
   }
 
-  @variant hover {
-    @apply z-10 !scale-100;
-
-    & ~ [class*="card-stack-item"] {
-      @apply z-[calc(10-var(--index))];
-    }
-
-    & + [class*="card-stack-item"] {
-      @apply !scale-[calc(1-var(--scale-pct))];
-
-      & + [class*="card-stack-item"] {
-        @apply !scale-[calc(1-2*var(--scale-pct))];
-
-        & ~ [class*="card-stack-item"] {
-          @apply !scale-[calc(1-3*var(--scale-pct))];
-        }
-      }
-    }
+  &::after {
+    @apply absolute inset-[1px] -z-1 rounded-(--radius) bg-(--bg) content-[''];
   }
-
-  &:last-child {
-    @apply scale-100;
-  }
-  &:has(+ :last-child) {
-    @apply scale-[calc(1-var(--scale-pct))];
-  }
-  &:has(+ * + :last-child) {
-    @apply scale-[calc(1-2*var(--scale-pct))];
-  }
-}
-
-@keyframes rotateSun {
-  100% {
-    transform: rotateY(360deg);
-  }
-}
-
-@keyframes moveSun {
-  0% {
-    transform: translateY(0);
-  }
-  50% {
-    transform: translateY(-20px);
-  }
-  100% {
-    transform: translateY(0);
-  }
-}
-
-.sun-wrapper {
-  transform-style: preserve-3d;
-  perspective: 100rem;
-  width: 30rem;
-  aspect-ratio: 1/1;
-  margin: auto;
-  &.parent-sun {
-    animation: moveSun 6s linear infinite;
-  }
-  *,
-  *:after,
-  *:before {
-    content: "";
-    position: absolute;
-    left: 0;
-    top: 0;
-    box-sizing: border-box;
-    transform-style: preserve-3d;
-  }
-}
-
-.sun {
-  &,
-  span,
-  &:after,
-  .rotate {
-    width: 30rem;
-    aspect-ratio: 1/1;
-    border-radius: 50%;
-    border: 1px solid rgba(242, 201, 97, 0.3);
-  }
-
-  span:nth-of-type(1) {
-    transform: rotateY(0deg) rotateX(0deg);
-  }
-  span:nth-of-type(2) {
-    transform: rotateY(20deg) rotateX(20deg);
-  }
-  span:nth-of-type(3) {
-    transform: rotateY(40deg) rotateX(40deg);
-  }
-  span:nth-of-type(4) {
-    transform: rotateY(60deg) rotateX(60deg);
-  }
-  span:nth-of-type(5) {
-    transform: rotateY(80deg) rotateX(80deg);
-  }
-  span:nth-of-type(6) {
-    transform: rotateY(100deg) rotateX(100deg);
-  }
-  span:nth-of-type(7) {
-    transform: rotateY(120deg) rotateX(120deg);
-  }
-  span:nth-of-type(8) {
-    transform: rotateY(140deg) rotateX(140deg);
-  }
-  span:nth-of-type(9) {
-    transform: rotateY(160deg) rotateX(160deg);
-  }
-  span:nth-of-type(10) {
-    transform: rotateY(180deg) rotateX(180deg);
-  }
-  span:nth-of-type(11) {
-    transform: rotateY(200deg) rotateX(200deg);
-  }
-  span:nth-of-type(12) {
-    transform: rotateY(220deg) rotateX(220deg);
-  }
-  span:nth-of-type(13) {
-    transform: rotateY(240deg) rotateX(240deg);
-  }
-  span:nth-of-type(14) {
-    transform: rotateY(260deg) rotateX(260deg);
-  }
-  span:nth-of-type(15) {
-    transform: rotateY(280deg) rotateX(280deg);
-  }
-  span:nth-of-type(16) {
-    transform: rotateY(300deg) rotateX(300deg);
-  }
-  span:nth-of-type(17) {
-    transform: rotateY(320deg) rotateX(320deg);
-  }
-  span:nth-of-type(18) {
-    transform: rotateY(340deg) rotateX(340deg);
-  }
-  span:nth-of-type(19) {
-    transform: rotateY(360deg) rotateX(360deg);
-  }
-  span:nth-of-type(20) {
-    transform: rotateY(380deg) rotateX(380deg);
-  }
-  span:nth-of-type(21) {
-    transform: rotateY(400deg) rotateX(400deg);
-  }
-  span:nth-of-type(22) {
-    transform: rotateY(420deg) rotateX(420deg);
-  }
-  span:nth-of-type(23) {
-    transform: rotateY(440deg) rotateX(440deg);
-  }
-  span:nth-of-type(24) {
-    transform: rotateY(460deg) rotateX(460deg);
-  }
-  span:nth-of-type(25) {
-    transform: rotateY(480deg) rotateX(480deg);
-  }
-  span:nth-of-type(26) {
-    transform: rotateY(500deg) rotateX(500deg);
-  }
-  span:nth-of-type(27) {
-    transform: rotateY(520deg) rotateX(520deg);
-  }
-  span:nth-of-type(28) {
-    transform: rotateY(540deg) rotateX(540deg);
-  }
-  span:nth-of-type(29) {
-    transform: rotateY(560deg) rotateX(560deg);
-  }
-  span:nth-of-type(30) {
-    transform: rotateY(580deg) rotateX(580deg);
-  }
-
-  .rotate {
-    perspective: none;
-    animation: rotateSun 15s linear infinite;
-  }
-
-  span {
-    box-shadow: 20px 0 50px rgba(242, 201, 97, 0.3);
-    /* filter: blur(.3rem) brightness(0.9); */
-  }
-}
-
-.sun:after {
-  background: radial-gradient(#f81c53aa, #ff9500, white);
-  transform: translate3d(0, 0, 20rem) scale(0.66);
-  filter: blur(1rem);
 }
