+++ components/icons/checker.tsx
@@ -1,22 +1,22 @@
 import Image from "next/image";
 import { HTMLAttributes } from "react";
 
-interface CheckerProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number;
-}
+import { cn } from "@/lib/utils";
 
 export default function Checker({
   className,
-  size = 32,
   ...props
-}: CheckerProps) {
+}: HTMLAttributes<HTMLDivElement>) {
   return (
-    <div className={className} {...props}>
+    <div
+      {...props}
+      className={cn("relative aspect-square shrink-0", className)}
+    >
       <Image
         src="/icons/checker.svg"
         alt="Checker Icon"
-        width={size}
-        height={size}
+        fill
+        className="object-contain object-center"
       />
     </div>
   );
