+++ components/icons/dataScientist.tsx
@@ -1,22 +1,22 @@
 import Image from "next/image";
 import { HTMLAttributes } from "react";
 
-interface DataScientistProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number;
-}
+import { cn } from "@/lib/utils";
 
 export default function DataScientist({
   className,
-  size = 32,
   ...props
-}: DataScientistProps) {
+}: HTMLAttributes<HTMLDivElement>) {
   return (
-    <div className={className} {...props}>
+    <div
+      {...props}
+      className={cn("relative aspect-square shrink-0", className)}
+    >
       <Image
         src="/icons/data-scientist.svg"
         alt="Data Scientist Icon"
-        width={size}
-        height={size}
+        fill
+        className="object-contain object-center"
       />
     </div>
   );
